# expression: implement vectorized evaluation for builtinGEIntSig