--- a/expression/builtin_compare_vec.go
+++ b/expression/builtin_compare_vec.go
@@ -164,11 +164,35 @@ func (b *builtinGreatestIntSig) vectorized() bool {
 }
 
 func (b *builtinGEIntSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinGEIntSig) vecEvalInt(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	var err error
+	var buf0, buf1 *chunk.Column
+	buf0, err = b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf0)
+	if err = b.args[0].VecEvalInt(b.ctx, input, buf0); err != nil {
+		return err
+	}
+	buf1, err = b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf1)
+	if err = b.args[1].VecEvalInt(b.ctx, input, buf1); err != nil {
+		return err
+	}
+
+	result.ResizeInt64(n, false)
+	vecCompareInt(mysql.HasUnsignedFlag(b.args[0].GetType().Flag), mysql.HasUnsignedFlag(b.args[1].GetType().Flag), buf0, buf1, result)
+	result.MergeNulls(buf0, buf1)
+	vecResOfGE(result.Int64s())
+	return nil
 }
 
 func (b *builtinLeastRealSig) vectorized() bool {
@@ -533,6 +557,17 @@ func vecResOfGT(res []int64) {
 	n := len(res)
 	for i := 0; i < n; i++ {
 		if res[i] > 0 {
+			res[i] = 1
+		} else {
+			res[i] = 0
+		}
+	}
+}
+
+func vecResOfGE(res []int64) {
+	n := len(res)
+	for i := 0; i < n; i++ {
+		if res[i] >= 0 {
 			res[i] = 1
 		} else {
 			res[i] = 0